# Simplify unpackString