Smaller, simpler, and (imperceptibly) faster by always doing everything in a single loop and keeping a bookmark rather than a `WriteByte` for every unescaped character:
```
$ go test -benchmem -bench=UnpackString >benchmark.new.txt && benchstat *old*txt *new*txt
name                      old time/op    new time/op    delta
UnpackString/Escaped-8       100ns ± 0%      96ns ± 0%   ~     (p=1.000 n=1+1)
UnpackString/Unescaped-8    55.1ns ± 0%    53.8ns ± 0%   ~     (p=1.000 n=1+1)

name                      old alloc/op   new alloc/op   delta
UnpackString/Escaped-8       32.0B ± 0%     32.0B ± 0%   ~     (all equal)
UnpackString/Unescaped-8     32.0B ± 0%     32.0B ± 0%   ~     (all equal)

name                      old allocs/op  new allocs/op  delta
UnpackString/Escaped-8        1.00 ± 0%      1.00 ± 0%   ~     (all equal)
UnpackString/Unescaped-8      1.00 ± 0%      1.00 ± 0%   ~     (all equal)
```