@miekg 2019-09-23
like it, also removes some of the duplication.