@tmthrgd 2019-09-23
This is nearly exactly the same as #1009 and is much more likely to over-allocate than #1011.