@miekg 2019-09-23
[ Quoting <notifications@github.com> in "Re: [miekg/dns] Simplify unpackStri..." ]
>That being said, I don't actually object to this.

merge?